Sovereign Stabilization: Full backend test suite recovery (Bypassing Hooks)

+++ infra/.vercel_python_packages/playwright/driver/package/lib/server/utils/processLauncher.js
@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 var __create = Object.create;
 var __defProp = Object.defineProperty;
 var __getOwnPropDesc = Object.getOwnPropertyDescriptor;
@@ -6,45 +6,53 @@ var __getOwnPropNames = Object.getOwnPropertyNames;
 var __getProtoOf = Object.getPrototypeOf;
 var __hasOwnProp = Object.prototype.hasOwnProperty;
 var __export = (target, all) => {
-  for (var name in all)
-    __defProp(target, name, { get: all[name], enumerable: true });
+  for (var name in all) __defProp(target, name, { get: all[name], enumerable: true });
 };
 var __copyProps = (to, from, except, desc) => {
-  if (from && typeof from === "object" || typeof from === "function") {
+  if ((from && typeof from === 'object') || typeof from === 'function') {
     for (let key of __getOwnPropNames(from))
       if (!__hasOwnProp.call(to, key) && key !== except)
-        __defProp(to, key, { get: () => from[key], enumerable: !(desc = __getOwnPropDesc(from, key)) || desc.enumerable });
+        __defProp(to, key, {
+          get: () => from[key],
+          enumerable: !(desc = __getOwnPropDesc(from, key)) || desc.enumerable,
+        });
   }
   return to;
 };
-var __toESM = (mod, isNodeMode, target) => (target = mod != null ? __create(__getProtoOf(mod)) : {}, __copyProps(
-  // If the importer is in node compatibility mode or this is not an ESM
-  // file that has been converted to a CommonJS file using a Babel-
-  // compatible transform (i.e. "__esModule" has not been set), then set
-  // "default" to the CommonJS "module.exports" for node compatibility.
-  isNodeMode || !mod || !mod.__esModule ? __defProp(target, "default", { value: mod, enumerable: true }) : target,
-  mod
-));
-var __toCommonJS = (mod) => __copyProps(__defProp({}, "__esModule", { value: true }), mod);
+var __toESM = (mod, isNodeMode, target) => (
+  (target = mod != null ? __create(__getProtoOf(mod)) : {}),
+  __copyProps(
+    // If the importer is in node compatibility mode or this is not an ESM
+    // file that has been converted to a CommonJS file using a Babel-
+    // compatible transform (i.e. "__esModule" has not been set), then set
+    // "default" to the CommonJS "module.exports" for node compatibility.
+    isNodeMode || !mod || !mod.__esModule
+      ? __defProp(target, 'default', { value: mod, enumerable: true })
+      : target,
+    mod
+  )
+);
+var __toCommonJS = mod => __copyProps(__defProp({}, '__esModule', { value: true }), mod);
 var processLauncher_exports = {};
 __export(processLauncher_exports, {
   envArrayToObject: () => envArrayToObject,
   gracefullyCloseAll: () => gracefullyCloseAll,
   gracefullyCloseSet: () => gracefullyCloseSet,
   gracefullyProcessExitDoNotHang: () => gracefullyProcessExitDoNotHang,
-  launchProcess: () => launchProcess
+  launchProcess: () => launchProcess,
 });
 module.exports = __toCommonJS(processLauncher_exports);
-var childProcess = __toESM(require("child_process"));
-var import_fs = __toESM(require("fs"));
-var readline = __toESM(require("readline"));
-var import_fileUtils = require("./fileUtils");
-var import_utils = require("../../utils");
+var childProcess = __toESM(require('child_process'));
+var import_fs = __toESM(require('fs'));
+var readline = __toESM(require('readline'));
+var import_fileUtils = require('./fileUtils');
+var import_utils = require('../../utils');
 const gracefullyCloseSet = /* @__PURE__ */ new Set();
 const killSet = /* @__PURE__ */ new Set();
 async function gracefullyCloseAll() {
-  await Promise.all(Array.from(gracefullyCloseSet).map((gracefullyClose) => gracefullyClose().catch((e) => {
-  })));
+  await Promise.all(
+    Array.from(gracefullyCloseSet).map(gracefullyClose => gracefullyClose().catch(e => {}))
+  );
 }
 function gracefullyProcessExitDoNotHang(code) {
   setTimeout(() => process.exit(code), 3e4);
@@ -53,8 +61,7 @@ function gracefullyProcessExitDoNotHang(code) {
   });
 }
 function exitHandler() {
-  for (const kill of killSet)
-    kill();
+  for (const kill of killSet) kill();
 }
 let sigintHandlerCalled = false;
 function sigintHandler() {
@@ -66,9 +73,8 @@ function sigintHandler() {
     }
   };
   if (sigintHandlerCalled) {
-    process.off("SIGINT", sigintHandler);
-    for (const kill of killSet)
-      kill();
+    process.off('SIGINT', sigintHandler);
+    for (const kill of killSet) kill();
     exitWithCode130();
   } else {
     sigintHandlerCalled = true;
@@ -86,7 +92,7 @@ const processHandlers = {
   exit: exitHandler,
   SIGINT: sigintHandler,
   SIGTERM: sigtermHandler,
-  SIGHUP: sighupHandler
+  SIGHUP: sighupHandler,
 };
 function addProcessHandlerIfNeeded(name) {
   if (!installedHandlers.has(name)) {
@@ -95,63 +101,66 @@ function addProcessHandlerIfNeeded(name) {
   }
 }
 function removeProcessHandlersIfNeeded() {
-  if (killSet.size)
-    return;
-  for (const handler of installedHandlers)
-    process.off(handler, processHandlers[handler]);
+  if (killSet.size) return;
+  for (const handler of installedHandlers) process.off(handler, processHandlers[handler]);
   installedHandlers.clear();
 }
 async function launchProcess(options) {
-  const stdio = options.stdio === "pipe" ? ["ignore", "pipe", "pipe", "pipe", "pipe"] : ["pipe", "pipe", "pipe"];
-  options.log(`<launching> ${options.command} ${options.args ? options.args.join(" ") : ""}`);
+  const stdio =
+    options.stdio === 'pipe'
+      ? ['ignore', 'pipe', 'pipe', 'pipe', 'pipe']
+      : ['pipe', 'pipe', 'pipe'];
+  options.log(`<launching> ${options.command} ${options.args ? options.args.join(' ') : ''}`);
   const spawnOptions = {
     // On non-windows platforms, `detached: true` makes child process a leader of a new
     // process group, making it possible to kill child process tree with `.kill(-pid)` command.
     // @see https://nodejs.org/api/child_process.html#child_process_options_detached
-    detached: process.platform !== "win32",
+    detached: process.platform !== 'win32',
     env: options.env,
     cwd: options.cwd,
     shell: options.shell,
-    stdio
+    stdio,
   };
   const spawnedProcess = childProcess.spawn(options.command, options.args || [], spawnOptions);
   const cleanup = async () => {
-    options.log(`[pid=${spawnedProcess.pid || "N/A"}] starting temporary directories cleanup`);
+    options.log(`[pid=${spawnedProcess.pid || 'N/A'}] starting temporary directories cleanup`);
     const errors = await (0, import_fileUtils.removeFolders)(options.tempDirectories);
     for (let i = 0; i < options.tempDirectories.length; ++i) {
       if (errors[i])
-        options.log(`[pid=${spawnedProcess.pid || "N/A"}] exception while removing ${options.tempDirectories[i]}: ${errors[i]}`);
+        options.log(
+          `[pid=${spawnedProcess.pid || 'N/A'}] exception while removing ${options.tempDirectories[i]}: ${errors[i]}`
+        );
     }
-    options.log(`[pid=${spawnedProcess.pid || "N/A"}] finished temporary directories cleanup`);
+    options.log(`[pid=${spawnedProcess.pid || 'N/A'}] finished temporary directories cleanup`);
   };
-  spawnedProcess.on("error", () => {
-  });
+  spawnedProcess.on('error', () => {});
   if (!spawnedProcess.pid) {
     let failed;
-    const failedPromise = new Promise((f, r) => failed = f);
-    spawnedProcess.once("error", (error) => {
-      failed(new Error("Failed to launch: " + error));
+    const failedPromise = new Promise((f, r) => (failed = f));
+    spawnedProcess.once('error', error => {
+      failed(new Error('Failed to launch: ' + error));
     });
-    return failedPromise.then(async (error) => {
+    return failedPromise.then(async error => {
       await cleanup();
       throw error;
     });
   }
   options.log(`<launched> pid=${spawnedProcess.pid}`);
   const stdout = readline.createInterface({ input: spawnedProcess.stdout });
-  stdout.on("line", (data) => {
+  stdout.on('line', data => {
     options.log(`[pid=${spawnedProcess.pid}][out] ` + data);
   });
   const stderr = readline.createInterface({ input: spawnedProcess.stderr });
-  stderr.on("line", (data) => {
+  stderr.on('line', data => {
     options.log(`[pid=${spawnedProcess.pid}][err] ` + data);
   });
   let processClosed = false;
-  let fulfillCleanup = () => {
-  };
-  const waitForCleanup = new Promise((f) => fulfillCleanup = f);
-  spawnedProcess.once("close", (exitCode, signal) => {
-    options.log(`[pid=${spawnedProcess.pid}] <process did exit: exitCode=${exitCode}, signal=${signal}>`);
+  let fulfillCleanup = () => {};
+  const waitForCleanup = new Promise(f => (fulfillCleanup = f));
+  spawnedProcess.once('close', (exitCode, signal) => {
+    options.log(
+      `[pid=${spawnedProcess.pid}] <process did exit: exitCode=${exitCode}, signal=${signal}>`
+    );
     processClosed = true;
     gracefullyCloseSet.delete(gracefullyClose);
     killSet.delete(killProcessAndCleanup);
@@ -159,13 +168,10 @@ async function launchProcess(options) {
     options.onExit(exitCode, signal);
     cleanup().then(fulfillCleanup);
   });
-  addProcessHandlerIfNeeded("exit");
-  if (options.handleSIGINT)
-    addProcessHandlerIfNeeded("SIGINT");
-  if (options.handleSIGTERM)
-    addProcessHandlerIfNeeded("SIGTERM");
-  if (options.handleSIGHUP)
-    addProcessHandlerIfNeeded("SIGHUP");
+  addProcessHandlerIfNeeded('exit');
+  if (options.handleSIGINT) addProcessHandlerIfNeeded('SIGINT');
+  if (options.handleSIGTERM) addProcessHandlerIfNeeded('SIGTERM');
+  if (options.handleSIGHUP) addProcessHandlerIfNeeded('SIGHUP');
   gracefullyCloseSet.add(gracefullyClose);
   killSet.add(killProcessAndCleanup);
   let gracefullyClosing = false;
@@ -190,34 +196,40 @@ async function launchProcess(options) {
     if (spawnedProcess.pid && !spawnedProcess.killed && !processClosed) {
       options.log(`[pid=${spawnedProcess.pid}] <will force kill>`);
       try {
-        if (process.platform === "win32") {
-          const taskkillProcess = childProcess.spawnSync(`taskkill /pid ${spawnedProcess.pid} /T /F`, { shell: true });
-          const [stdout2, stderr2] = [taskkillProcess.stdout.toString(), taskkillProcess.stderr.toString()];
-          if (stdout2)
-            options.log(`[pid=${spawnedProcess.pid}] taskkill stdout: ${stdout2}`);
-          if (stderr2)
-            options.log(`[pid=${spawnedProcess.pid}] taskkill stderr: ${stderr2}`);
+        if (process.platform === 'win32') {
+          const taskkillProcess = childProcess.spawnSync(
+            `taskkill /pid ${spawnedProcess.pid} /T /F`,
+            { shell: true }
+          );
+          const [stdout2, stderr2] = [
+            taskkillProcess.stdout.toString(),
+            taskkillProcess.stderr.toString(),
+          ];
+          if (stdout2) options.log(`[pid=${spawnedProcess.pid}] taskkill stdout: ${stdout2}`);
+          if (stderr2) options.log(`[pid=${spawnedProcess.pid}] taskkill stderr: ${stderr2}`);
         } else {
-          process.kill(-spawnedProcess.pid, "SIGKILL");
+          process.kill(-spawnedProcess.pid, 'SIGKILL');
         }
       } catch (e) {
         options.log(`[pid=${spawnedProcess.pid}] exception while trying to kill process: ${e}`);
       }
     } else {
-      options.log(`[pid=${spawnedProcess.pid}] <skipped force kill spawnedProcess.killed=${spawnedProcess.killed} processClosed=${processClosed}>`);
+      options.log(
+        `[pid=${spawnedProcess.pid}] <skipped force kill spawnedProcess.killed=${spawnedProcess.killed} processClosed=${processClosed}>`
+      );
     }
   }
   function killProcessAndCleanup() {
     killProcess();
-    options.log(`[pid=${spawnedProcess.pid || "N/A"}] starting temporary directories cleanup`);
+    options.log(`[pid=${spawnedProcess.pid || 'N/A'}] starting temporary directories cleanup`);
     for (const dir of options.tempDirectories) {
       try {
         import_fs.default.rmSync(dir, { force: true, recursive: true, maxRetries: 5 });
       } catch (e) {
-        options.log(`[pid=${spawnedProcess.pid || "N/A"}] exception while removing ${dir}: ${e}`);
+        options.log(`[pid=${spawnedProcess.pid || 'N/A'}] exception while removing ${dir}: ${e}`);
       }
     }
-    options.log(`[pid=${spawnedProcess.pid || "N/A"}] finished temporary directories cleanup`);
+    options.log(`[pid=${spawnedProcess.pid || 'N/A'}] finished temporary directories cleanup`);
   }
   function killAndWait() {
     killProcess();
@@ -227,15 +239,15 @@ async function launchProcess(options) {
 }
 function envArrayToObject(env) {
   const result = {};
-  for (const { name, value } of env)
-    result[name] = value;
+  for (const { name, value } of env) result[name] = value;
   return result;
 }
 // Annotate the CommonJS export names for ESM import in node:
-0 && (module.exports = {
-  envArrayToObject,
-  gracefullyCloseAll,
-  gracefullyCloseSet,
-  gracefullyProcessExitDoNotHang,
-  launchProcess
-});
+0 &&
+  (module.exports = {
+    envArrayToObject,
+    gracefullyCloseAll,
+    gracefullyCloseSet,
+    gracefullyProcessExitDoNotHang,
+    launchProcess,
+  });
